backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
python-multipart==0.0.6
pdfkit==1.0.0
wkhtmltopdf==0.2
aiofiles==23.2.1
email-validator==2.1.0